refactor(all files): adding suffix of respective item
