+++ src/config/gameOptionsConfig.ts
